requirements.txt

flask
pdf2image
Pillow
pypdf2==2.12.1
requests
